pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.1"
description = "An example Xpresso app on GKE"
authors = [
    "Adrian Garcia Badaracco <1755071+adriangb@users.noreply.github.com>",
]
license = "MIT"
packages = [{ include = "app", from = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
xpresso = "~0"  # unpinned for now
uvicorn = { extras = ["standard"], version = "~0.17" }
asyncpg = "^0.25.0"
python-json-logger = "^2.0.2"
orjson = "^3.6.7"
argon2-cffi = "^21.3.0"
migri = {extras = ["postgresql"], version = "^0.6.3"}
pydantic = {extras = ["email"], version = "^1.9.0"}
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]
requests = "~2"     # for TestClient
black = {version = "*", allow-prereleases = true}
pytest = "~7"
pytest-sugar = "~0"
pre-commit = "~2"
isort = "~5"
flake8 = "~4"
mypy = ">=0.931"
pytest-cov = "~3"
coverage = {extras = ["toml"], version = "^~6"}
httpx = "^0.22.0"
asgi-lifespan = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
known_first_party = ["app", "tests"]

[tool.coverage.run]
branch = true
source = ["app/app", "app/tests"]
dynamic_context = "test_function"

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    '\.\.\.',
    "raise NotImplementedError",
    "TYPE_CHECKING",
    "__repr__",
    "import",
]

[tool.pytest.ini_options]
testpaths = ["app/tests"]

[tool.mypy]
files = "app/**/*.py"
strict = true
warn_unused_ignores = false
